exclude.snps {ARTP2}	R Documentation

Excluding SNPs from specified regions

Description

exclude.snps is used to exclude SNPs from specified regions, for example, the GWAS hits and their neighbors.

Usage

exclude.snps(stat, excluded.regions)

Arguments

stat
a data frame containing at least two columns Chr and Pos, i.e., the chromosome numbers and base-pair position (bp units) of SNPs.

excluded.regions
a data frame specifying the regions to be excluded. It must contains columns Chr, Start, and End. The unit is base-pair (bp). SNPs within [Start, End] will be excluded.

Value

A data frame containing a subset of stat after excluding SNPs in specified regions.

[Package ARTP2 version 1.2.32 Index]