geno {ARTP2}	R Documentation

A vector of file names used in example of rARTP.

Description

Names of 15 genotype files. Those files are plain text files, each stores genotypes data of a gene.

Usage

data(geno)

Examples

  data(geno)
  head(geno)

[Package ARTP2 version 1.2.32 Index]