inflation {ARTP2}	R Documentation

Computing the genomic control inflation factor from a given numeric vector

Description

This function gives the inflation of a set of p-values (or 1-df chi-squared statistics).

Usage

inflation(p, is.p, na.rm = FALSE)

Arguments

p
a numeric vector. It can contain p-values or 1-df chi-squared statistics.

is.p
a logical value indicating whether p contain p-values or a vector of 1-df chi-squared statistics.

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

Value

The inflation factor is returned.

See Also

meta

Examples

set.seed(0)
p <- runif(10)
inflation(p, TRUE)

[Package ARTP2 version 1.2.32 Index]