meta {ARTP2}	R Documentation

Single-marker meta-analysis

Description

Perform the single-marker meta-analysis with inflation adjustment. The inverse-variance weighting approach is used.

Usage

  meta(summary.files, lambda = NULL, sel.snps = NULL, only.meta=TRUE, ambig.by.AF=FALSE)

Arguments

summary.files
a character vector of file names containing the summary results of SNPs included in one or multiple studies. Each file must be able to be read by read.table. Each file must have columns called "SNP", "RefAllele", "EffectAllele", "BETA", and at least one of "SE", "P".

lambda
a numeric vector of inflation factors. Each file in summary.files should have one inflation factor specified in lambda. NULL if inflation is not adjusted.

sel.snps
a character vector of SNPs to be used in meta-analysis. The default is NULL, i.e., all SNPs are used.

only.meta
TRUE if do not returned individual summary data. The default is TRUE.

ambig.by.AF
TRUE or FALSE to align SNPs with ambiguous alleles by allele frequency (see options). The default is FALSE.

Details

The inverse-variance method is used by assuming a fixed effect model. The standard error is rescaled by sqrt{lambda}.

Value

meta returns a list containing

meta.stat
a data frame of summary statistics from meta-analysis. The summary statistics of individual studies specified in summary.files are also returned in the data frame if only.meta is FALSE. The standard error of individual studies are rescaled by sqrt{lambda}.

conf.snps
a character vector of SNPs with conflictive allele information.

Examples

study1 <- system.file("extdata", package = "ARTP2", "study1.txt.gz")
study2 <- system.file("extdata", package = "ARTP2", "study2.txt.gz")
snps <- c('rs13266821', 'rs4824130', 'rs1792438', 'rs1000047', 
          'rs1000017', 'rs6066771', 'rs12508128')

m1 <- meta(summary.files = c(study1, study2), lambda = c(1.10, 1.08), 
           sel.snps = snps)
m2 <- meta(summary.files = c(study1, study2), lambda = c(1.10, 1.08), 
           sel.snps = snps, only.meta=FALSE)

m1$conf.snps

m1$meta.stat
m2$meta.stat

[Package ARTP2 version 1.2.32 Index]