ref.geno {ARTP2}	R Documentation

A dataset used in example of sARTP.

Description

A data frame ref.geno with genotypes of 503 observations on 2654 SNPs. The genotypes are encoded by SNPs' minor and major alleles so that this data frame can be used as reference of function sARTP.

Usage

data(ref.geno)

Examples

  data(ref.geno)
  head(ref.geno[, 1:5])

[Package ARTP2 version 1.2.32 Index]