sARTP.multiPop.SNP.Gene {ARTP2}	R Documentation

ARTP test for multiple independent populations

Description

Calculate gene and pathway p-values using the ARTP test and summary data from multiple independent populations.

Usage

sARTP.multiPop.SNP.Gene(summary.files.list, pathway, family, reference.list,
        lambda.list, ncases.list, ncontrols.list, nsamples.list, options.list=NULL, 
        options.merged.SNP=NULL, options.merged.gene=NULL, options.merged=NULL)

Arguments

summary.files.list
A list of summary files for all populations. Each element of the list should correspond to the summary files for one population. See the documentation for summary.files in sARTP.

pathway
See the documentation for pathway in sARTP.

family
A string taking values of 'gaussian' or 'binomial'.

reference.list
A list of reference sets for all populations. Each element of the list should correspond to the reference set for one population. See the documentation for reference in sARTP.

lambda.list
A list of lambda values for all populations. Each element of the list should correspond to the lambda values for one population. See the documentation for lambda in sARTP.

ncases.list
NULL or a list of the number of cases for all populations. Each element of the list should correspond to the number of cases for one population. See the documentation for ncases in sARTP.

ncontrols.list
NULL or a list of the number of controls for all populations. Each element of the list should correspond to the number of controls for one population. See the documentation for ncontrols in sARTP.

nsamples.list
NULL or a list of the number of samples for all populations. Each element of the list should correspond to the number of samples for one population. See the documentation for nsamples in sARTP.

options.list
NULL or a list of options for all populations. Each element of the list should correspond to the options for one population. The options for each population are mainly for SNP filtering, and each population can have different filtering criteria. See options.

options.merged.SNP
NULL or a list of options for the merged population in function sARTP.multiPop.SNP. This list of options corresponds to options.merged in sARTP.multiPop.SNP. See options. One specific option for sARTP.multiPop.SNP is merged.snp.method.

options.merged.gene
NULL or a list of options for the function warm.start.multiPop.gene. Currently, the available options are method and nthread. By default, the options are
list(method="FISHER", nthread=1, use.ranks.pathway=TRUE).

options.merged
NULL or a list of options for the merged population. See options. One specific option is merged.gene.method which gives the method of merging the gene p-values obtained from sARTP.multiPop.SNP and from warm.start.multiPop.gene. The possible values for merged.gene.method are "ACAT", "FISHER", and "MAX", with "MAX" being the default.

Details

This function calls both sARTP.multiPop.SNP and warm.start.multiPop.gene to obtain permuted gene p-values. The permuted gene p-values are then merged together to obtain the final gene p-values and the pathway p-value.

Internally, the expand.fold option is set to 0.

Value

A list containing the following components:

pathway.pvalue
final pathway p-value accounting for multiple comparisons.

gene.pvalue
a data frame containing gene name, number of SNPs in the gene that were included in the analysis, chromosome name, and the p-value for the gene accounting for multiple comparisons.

accurate
TRUE if options$nperm is large enougth to accurately estimate p-values, i.e., if the criteria
(sqrt(pvalue*(1-pvalue)/pathway.nperm)/pvalue < 0.1) &&
(ceiling(-log10(minimum gene pvalue)) < log10(nperm))
is satisfied.

References

Zhang H, Wheeler W, Hyland LP, Yang Y, Shi J, Chatterjee N, Yu K. (2016) A powerful procedure for pathway-based meta-analysis using summary statistics identifies 43 pathways associated with type II diabetes in European populations. PLoS Genetics 12(6): e1006122

Yu K, Li Q, Bergen AW, Pfeiffer RM, Rosenberg PS, Caporaso N, Kraft P, Chatterjee N. (2009) Pathway analysis by adaptive combination of P-values. Genet Epidemiol 33(8): 700 - 709

Zhang H, Shi J, Liang F, Wheeler W, Stolzenberg-Solomon R, Yu K. (2014) A fast multilocus test with adaptive SNP selection for large-scale genetic association studies. European Journal of Human Genetics 22: 696 - 702

See Also

options, sARTP, warm.start.multiPop.SNP, sARTP.multiPop.SNP.union_intersect

Examples

  library(ARTP2)

  # Pathway definition
  pathway <- system.file("extdata", package = "ARTP2", "pathway.txt.gz")

  # logistic regression is used in base model
  family <- 'binomial'

  ## Create data frame containing paths of PLINK files that are going to used as the reference
  ## As a small example, use chromosomes 20 and 22
  chr  <- c(20, 22)
  nchr <- length(chr)

  # Define reference data for each population
  fam <- vector("character", nchr)
  bim <- vector("character", nchr)
  bed <- vector("character", nchr)
  for(i in 1:nchr){
    fam[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".fam", sep = ""))
    bim[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".bim", sep = ""))
    bed[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".bed", sep = ""))
  }
  reference1 <- data.frame(fam, bim, bed, stringsAsFactors = FALSE)
  for(i in 1:nchr){
    fam[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.fam", sep = ""))
    bim[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.bim", sep = ""))
    bed[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.bed", sep = ""))
  }
  reference2 <- data.frame(fam, bim, bed, stringsAsFactors = FALSE)

  ## Other objects for population 1
  study1     <- system.file("extdata", package = "ARTP2", "study1.txt.gz")
  ncases1    <- list(c(9580, 2591))
  ncontrols1 <- list(c(53810, 3052))
  lambda1    <- 1.10
  options1   <- list(maf = .01, HWE.p = 1e-6, 
                     id.str = "Pop1", out.dir=getwd(), 
                     save.setup = FALSE)

  # For population 2
  study2     <- system.file("extdata", package = "ARTP2", "study_EAS.txt.gz")
  ncases2    <- list(1000)
  ncontrols2 <- list(1000)
  lambda2    <- 1.13
  options2   <- list(maf = .01, HWE.p = 1e-6,  
                     id.str = "Pop2", out.dir=getwd(), 
                     save.setup = FALSE)

  #ret <- sARTP.multiPop.SNP.Gene(list(study1, study2), pathway, family, 
  #                      list(reference1, reference2), 
  #                      list(lambda1, lambda2), ncases.list=list(ncases1, ncases2), 
  #                      ncontrols.list=list(ncontrols1, ncontrols2), 
  #                      options.list=list(options1, options2))

[Package ARTP2 version 1.2.32 Index]