sARTP.multiPop.SNP.union {ARTP2}	R Documentation

ARTP test for multiple independent populations

Description

Calculate gene and pathway p-values using the ARTP test and summary data from multiple independent populations.

Usage

sARTP.multiPop.SNP.union(summary.files.list, pathway, family, reference.list, lambda.list, 
      ncases.list, ncontrols.list, nsamples.list, options.list=NULL, options.merged=NULL)

Arguments

summary.files.list
A list of summary files for all populations. Each element of the list should correspond to the summary files for one population. See the documentation for summary.files in sARTP.

pathway
See the documentation for pathway in sARTP.

family
A string taking values of 'gaussian' or 'binomial'.

reference.list
A list of reference sets for all populations. Each element of the list should correspond to the reference set for one population. See the documentation for reference in sARTP.

lambda.list
A list of lambda values for all populations. Each element of the list should correspond to the lambda values for one population. See the documentation for lambda in sARTP.

ncases.list
NULL or a list of the number of cases for all populations. Each element of the list should correspond to the number of cases for one population. See the documentation for ncases in sARTP.

ncontrols.list
NULL or a list of the number of controls for all populations. Each element of the list should correspond to the number of controls for one population. See the documentation for ncontrols in sARTP.

nsamples.list
NULL or a list of the number of samples for all populations. Each element of the list should correspond to the number of samples for one population. See the documentation for nsamples in sARTP.

options.list
NULL or a list of options for all populations. Each element of the list should correspond to the options for one population. See options.

options.merged
NULL or a list of options for the merged population. See options.

Details

This function is similar to sARTP except that it deals with multiple populations instead of just a single population. All the populations should be independent. The algorithm for this function is as follows:
1. Loop over each population and perform the SNP/gene filtering based on each population's reference set and options.
2. Define the pathway to be the union of all SNPs remaining after the filtering in step 1.
3. Define the groups from the pathway defined in step 2.
4. For each population, recover the summary statistics (score and covariance matrix) within each group.
5. Merge the summary statistics together across all populations for each group.
6. Generate permuted test statistics for the merged population.
7. Compute the gene and pathway p-values for the merged population.

Handwritten notes on the algorithm are at .../inst/doc/SNP_based_pathway_union.pdf

Value

A list containing the following components:

pathway.pvalue
final pathway p-value accounting for multiple comparisons.

gene.pvalue
a data frame containing gene name, number of SNPs in the gene that were included in the analysis, chromosome name, and the p-value for the gene accounting for multiple comparisons.

pathway
a data frame defining the pathway that was actually tested after various filters applied.

model
a list containing detailed information of selected SNPs in each gene.

most.sig.genes
a character vector of genes selected by ARTP2. They are the most promising candidates, although their statistical significance is not guaranteed.

options
a list of options used in the analysis. See options

pathway.nperm
The number of permutations for computing the pathway p-value. This takes into account the option expand.fold.

accurate
TRUE if options$nperm is large enougth to accurately estimate p-values, i.e., if the criteria
(sqrt(pvalue*(1-pvalue)/pathway.nperm)/pvalue < 0.1) &&
(ceiling(-log10(minimum gene pvalue)) < log10(nperm))
is satisfied.

setup
a list containing necessary input for warm.start.multiPop.SNP.union. It can be written to a file by using the function save, then its path can be the input of warm.start.multiPop.SNP.union.

References

Zhang H, Wheeler W, Hyland LP, Yang Y, Shi J, Chatterjee N, Yu K. (2016) A powerful procedure for pathway-based meta-analysis using summary statistics identifies 43 pathways associated with type II diabetes in European populations. PLoS Genetics 12(6): e1006122

Yu K, Li Q, Bergen AW, Pfeiffer RM, Rosenberg PS, Caporaso N, Kraft P, Chatterjee N. (2009) Pathway analysis by adaptive combination of P-values. Genet Epidemiol 33(8): 700 - 709

Zhang H, Shi J, Liang F, Wheeler W, Stolzenberg-Solomon R, Yu K. (2014) A fast multilocus test with adaptive SNP selection for large-scale genetic association studies. European Journal of Human Genetics 22: 696 - 702

See Also

options, sARTP, warm.start.multiPop.SNP.union

Examples

  library(ARTP2)

  # Pathway definition
  pathway <- system.file("extdata", package = "ARTP2", "pathway.txt.gz")

  # logistic regression is used in base model
  family <- 'binomial'

  ## Create data frame containing paths of PLINK files that are going to used as the reference
  ## As a small example, use chromosomes 20 and 22
  chr  <- c(20, 22)
  nchr <- length(chr)

  # Define reference data for each population
  fam <- vector("character", nchr)
  bim <- vector("character", nchr)
  bed <- vector("character", nchr)
  for(i in 1:nchr){
    fam[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".fam", sep = ""))
    bim[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".bim", sep = ""))
    bed[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".bed", sep = ""))
  }
  reference1 <- data.frame(fam, bim, bed, stringsAsFactors = FALSE)
  for(i in 1:nchr){
    fam[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.fam", sep = ""))
    bim[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.bim", sep = ""))
    bed[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.bed", sep = ""))
  }
  reference2 <- data.frame(fam, bim, bed, stringsAsFactors = FALSE)

  ## Other objects for population 1
  study1     <- system.file("extdata", package = "ARTP2", "study1.txt.gz")
  ncases1    <- list(c(9580, 2591))
  ncontrols1 <- list(c(53810, 3052))
  lambda1    <- 1.10
  options1   <- list(maf = .01, HWE.p = 1e-6, 
                     id.str = "Pop1", out.dir=getwd(), 
                     save.setup = FALSE)

  # For population 2
  study2     <- system.file("extdata", package = "ARTP2", "study_EAS.txt.gz")
  ncases2    <- list(1000)
  ncontrols2 <- list(1000)
  lambda2    <- 1.13
  options2   <- list(maf = .01, HWE.p = 1e-6,  
                     id.str = "Pop2", out.dir=getwd(), 
                     save.setup = FALSE)

  #ret <- sARTP.multiPop.SNP.union(list(study1, study2), pathway, family, list(reference1, reference2), 
  #                      list(lambda1, lambda2), ncases.list=list(ncases1, ncases2), 
  #                      ncontrols.list=list(ncontrols1, ncontrols2), 
  #                      options.list=list(options1, options2), options.merged=NULL)

[Package ARTP2 version 1.2.32 Index]