warm.start.multiPop.SNP {ARTP2}	R Documentation

Performing ARTP2 test with warm start for multiple independent populations

Description

This function is designed to accelerate the ARTP2 test in practice. It uses pre-calculated and reusable statistics as input and allow the users to try different testing configuration more efficiently. See Details for more information.

Usage

warm.start.multiPop.SNP(setup, nperm = NULL, lambda = 1.0, nthread = NULL)

Arguments

setup
an R object created by sARTP.multiPop.SNP. It is a list containing necessary statistics for computing p-values.

nperm
the number of permutations. If it is NULL as default, then the value in the saved setup will be used. See Details.

lambda
inflation factor to be adjusted in pathway analysis. lambda in this function can only be a single numeric number, which is different from the one in
sARTP.multiPop.SNP.union_intersect. The default is 1.0.

nthread
number of threads to be used in permutation. NULL if setup$options$nthread is used.

Details

An ARTP2 test has two major steps in testing an association. The first step applys data clean criteria and creates necessary and reusable statistics, which can be time-consuming for large pathways. The second step performs the testing procedure to estimate the pathway or gene-level p-value. warm.start.multiPop.SNP.union_intersect focuses on the second step.

The first step can be done by using sARTP.multiPop.SNP if options$only.setup is set to TRUE. The output object, setup, can be used as the first argument of
warm.start.multiPop.SNP.union_intersect. With warm.start.multiPop.SNP, users can try different configurations to perform various tests allowed by the ARTP2 framework, but avoid long waiting time for data cleaning. Commonly used options in setup$options include method, inspect.snp.n, inspect.gene.n, nperm, etc.

Note that sARTP.multiPop.SNP can produce the final p-value directly if options$only.setup is FALSE.

The setup is supposed to have all components defined in sARTP. If nperm is NULL, then it will be set as setup$options$nperm. The users can also pass lambda if a second round genomic control is needed. However, unlike in sARTP.multiPop.SNP, lambda here can only be a single numeric number rather than a vector. Options nperm and lambda are the most useful ones in using warm.start.multiPop.SNP so we highlight them in the interface. Users can modify any option values in setup$options directly to get more controls of the testing procedure. See options for more details about how to set setup$options.

Except for setup$options, all other components in setup should not be modified by users.

Value

A list containing the following components:

pathway.pvalue
final pathway p-value accounting for multiple comparisons.

gene.pvalue
a data frame containing gene name, number of SNPs in the gene that were included in the analysis, chromosome name, and the p-value for the gene accounting for multiple comparisons.

pathway
a data frame defining the pathway that was actually tested after various filters applied.

model
a list containing detailed information of selected SNPs in each gene.

most.sig.genes
a character vector of genes selected by ARTP2. They are the most promising candidates, although their statistical significance is not guaranteed.

options
a list of options used in the analysis. See options

test.timing
timing information (in sec)

pathway.nperm
The number of permutations for computing the pathway p-value. This takes into account the option expand.fold.

accurate
TRUE if options$nperm is large enougth to accurately estimate p-values, i.e., if the criteria
(sqrt(pvalue*(1-pvalue)/pathway.nperm)/pvalue < 0.1) &&
(ceiling(-log10(minimum gene pvalue)) < log10(nperm))
is satisfied.

References

Zhang H, Wheeler W, Hyland LP, Yang Y, Shi J, Chatterjee N, Yu K. (2016) A powerful procedure for pathway-based meta-analysis using summary statistics identifies 43 pathways associated with type II diabetes in European populations. PLoS Genetics 12(6): e1006122

Yu K, Li Q, Bergen AW, Pfeiffer RM, Rosenberg PS, Caporaso N, Kraft P, Chatterjee N. (2009) Pathway analysis by adaptive combination of P-values. Genet Epidemiol 33(8): 700 - 709

Zhang H, Shi J, Liang F, Wheeler W, Stolzenberg-Solomon R, Yu K. (2014) A fast multilocus test with adaptive SNP selection for large-scale genetic association studies. European Journal of Human Genetics: 22, 696 - 702

See Also

sARTP.multiPop.SNP, options

Examples

  library(ARTP2)

  # Pathway definition
  pathway <- system.file("extdata", package = "ARTP2", "pathway.txt.gz")

  # logistic regression is used in base model
  family <- 'binomial'

  ## Create data frame containing paths of PLINK files that are going to used as the reference
  ## As a small example, use chromosomes 20 and 22
  chr  <- c(20, 22)
  nchr <- length(chr)

  # Define reference data for each population
  fam <- vector("character", nchr)
  bim <- vector("character", nchr)
  bed <- vector("character", nchr)
  for(i in 1:nchr){
    fam[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".fam", sep = ""))
    bim[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".bim", sep = ""))
    bed[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".bed", sep = ""))
  }
  reference1 <- data.frame(fam, bim, bed, stringsAsFactors = FALSE)
  for(i in 1:nchr){
    fam[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.fam", sep = ""))
    bim[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.bim", sep = ""))
    bed[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], "_EAS.bed", sep = ""))
  }
  reference2 <- data.frame(fam, bim, bed, stringsAsFactors = FALSE)

  ## Other objects for population 1
  study1     <- system.file("extdata", package = "ARTP2", "study1.txt.gz")
  ncases1    <- list(c(9580, 2591))
  ncontrols1 <- list(c(53810, 3052))
  lambda1    <- 1.10
  options1   <- list(maf = .01, HWE.p = 1e-6, 
                     id.str = "Pop1", out.dir=getwd(), 
                     save.setup = FALSE)

  # For population 2
  study2     <- system.file("extdata", package = "ARTP2", "study_EAS.txt.gz")
  ncases2    <- list(1000)
  ncontrols2 <- list(1000)
  lambda2    <- 1.13
  options2   <- list(maf = .01, HWE.p = 1e-6,  
                     id.str = "Pop2", out.dir=getwd(), 
                     save.setup = FALSE)

  #setup <- sARTP.multiPop.SNP(list(study1, study2), pathway, family, list(reference1, reference2), 
  #                      list(lambda1, lambda2), ncases.list=list(ncases1, ncases2), 
  #                      ncontrols.list=list(ncontrols1, ncontrols2), 
  #                      options.list=list(options1, options2),
  #                      options.merged=list(only.setup=TRUE))

  #warm.start.multiPop.SNP(setup)

[Package ARTP2 version 1.2.32 Index]