warm.start.multiPop.gene {ARTP2}	R Documentation

ARTP2 for multiple independent populations

Description

To perform the ARTP2 test by merging gene results from multiple independent populations.

Usage

warm.start.multiPop.gene(sARTP2.list, method="FISHER", nthread=1, delete.gene.files=TRUE,
                         expand.fold=0, use.ranks.pathway=TRUE)

Arguments

sARTP2.list
A list of return objects from sARTP or .rda files containing the returned objects from sARTP for each population.

method
One of "MAX", "META", "FISHER", "ACAT" or "WGTFISHER" for the method to combine the permuted test statistics from each population. The default is "FISHER".

nthread
Number of threads to be used. The default is 1.

delete.gene.files
TRUE or FALSE to delete the temporary gene files from each population. The default is TRUE.

expand.fold
The number of copies of the nperm permutations to expand. Setting this option to a positive integer will expand the number of permutations by randomly permuting the nperm permutations within each gene expand.fold times when computing the pathway p-value. The pathway p-value is then essentially computed using (1 + expand.fold)*nperm permutations. The default is 0.

use.ranks.pathway
TRUE or FALSE to transform the merged gene p-values into permutation p-values based on their ranks. This option only affects the pathway p-value and is only applied when expand.fold = 0. The default is TRUE.

Details

This function combines the gene p-values from the populations using Fisher's method for all permutations, where the number of permutations is the minimum from the populations. Note that if more permutations are needed, then each population must be first run with a greater number of permutations. If a particular gene is only present in a subset of the populations, then the combined gene p-value is from the subset of populations.

When sARTP is called for each population, the option options$delete.gene.files must be set to FALSE, and a different seed should be passed in using options$seed.

Value

A list containing the following components:

pathway.pvalue
final pathway p-value accounting for multiple comparisons.

gene.pvalue
a data frame containing gene name, number of SNPs in the gene that were included in the analysis, chromosome name, and the p-value for the gene accounting for multiple comparisons.

pathway
a data frame defining the pathway that was actually tested after various filters applied.

most.sig.genes
a character vector of genes selected by ARTP2. They are the most promising candidates, although their statistical significance is not guaranteed.

test.timing
timing information (in sec)

pathway.nperm
The number of permutations for computing the pathway p-value. This takes into account the option expand.fold.

accurate
TRUE if options$nperm is large enougth to accurately estimate p-values, i.e., if the criteria
(sqrt(pvalue*(1-pvalue)/pathway.nperm)/pvalue < 0.1) &&
(ceiling(-log10(minimum gene pvalue)) < log10(nperm))
is satisfied.

References

Zhang H, Wheeler W, Hyland LP, Yang Y, Shi J, Chatterjee N, Yu K. (2016) A powerful procedure for pathway-based meta-analysis using summary statistics identifies 43 pathways associated with type II diabetes in European populations. PLoS Genetics 12(6): e1006122

Yu K, Li Q, Bergen AW, Pfeiffer RM, Rosenberg PS, Caporaso N, Kraft P, Chatterjee N. (2009) Pathway analysis by adaptive combination of P-values. Genet Epidemiol 33(8): 700 - 709

Zhang H, Shi J, Liang F, Wheeler W, Stolzenberg-Solomon R, Yu K. (2014) A fast multilocus test with adaptive SNP selection for large-scale genetic association studies. European Journal of Human Genetics: 22, 696 - 702

See Also

sARTP, options

Examples

  library(ARTP2)
  ## Create data frame containing paths of build-in PLINK files that are going to used as reference
  ## As an example, use chromosomes 20-22
  chr  <- 20:22
  nchr <- length(chr)

  # Define reference data
  fam <- vector("character", nchr)
  bim <- vector("character", nchr)
  bed <- vector("character", nchr)
  for(i in 1:nchr){
    fam[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".fam", sep = ""))
    bim[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".bim", sep = ""))
    bed[i] <- system.file("extdata", package = "ARTP2", paste("chr", chr[i], ".bed", sep = ""))
  }
  reference <- data.frame(fam, bim, bed, stringsAsFactors = FALSE)

  # Pathways do not need to be the same for each population. 
  # For this example, the pathway will be the same
  pathway <- system.file("extdata", package = "ARTP2", "pathway.txt.gz")

  # logistic regression is used in base model, thus ncases and ncontrols should be specified. 
  family <- 'binomial'

  ## Objects for population 1
  study1     <- system.file("extdata", package = "ARTP2", "study1.txt.gz")
  ncases1    <- list(c(9580, 2591))
  ncontrols1 <- list(c(53810, 3052))
  lambda1    <- 1.10

  # Set the options for population 1. 
  #############################################
  # !!! delete.gene.files must be set to FALSE
  #############################################
  options1 <- list(inspect.snp.n = 2, nperm = 1e3, nthread=1,
                maf = .01, HWE.p = 1e-6,  gene.R2 = .9, 
                id.str = "Pop1", out.dir=getwd(), 
                save.setup = FALSE, 
                delete.gene.files=FALSE, seed=123)

  #ret1 <- sARTP(summary.files = study1, pathway, family, reference, lambda1, 
  #            ncases1, ncontrols1, options = options1)
 
  # For population 2
  study2     <- system.file("extdata", package = "ARTP2", "study2.txt.gz")
  ncases2    <- list(7638)
  ncontrols2 <- list(54319)
  lambda2    <- 1.08

  # Set the options for population 2. 
  #############################################
  # !!! delete.gene.files must be set to FALSE
  # A different seed is used
  #############################################
  options2 <- list(inspect.snp.n = 2, nperm = 1e3, nthread=1,
                maf = .01, HWE.p = 1e-6,  gene.R2 = .9, 
                id.str = "Pop2", out.dir=getwd(), 
                save.setup = FALSE, 
                delete.gene.files=FALSE, seed=456)

  #ret2 <- sARTP(summary.files = study2, pathway, family, reference, lambda2, 
  #             ncases2, ncontrols2, options = options2)

  # Run ARTP2 by merging population results
  #mrg <- warm.start.multiPop.gene(list(ret1, ret2))
  #mrg$pathway.pvalue
  #mrg$gene.pvalue

[Package ARTP2 version 1.2.32 Index]